+++ modules/canvaskit/npm_build/types/index.d.ts
@@ -457,6 +457,7 @@ export interface CanvasKit {
     readonly Path: PathConstructorAndFactory;
     readonly PictureRecorder: DefaultConstructor<PictureRecorder>;
     readonly TextStyle: TextStyleConstructor;
+    readonly SlottableTextProperty: SlottableTextPropertyConstructor;
 
     // Factories, i.e. things made with CanvasKit.Foo.MakeTurboEncabulator()
     readonly ParagraphBuilder: ParagraphBuilderFactory;
@@ -559,6 +560,10 @@ export interface CanvasKit {
     readonly TextBaseline: TextBaselineEnumValues;
     readonly TextDirection: TextDirectionEnumValues;
     readonly TextHeightBehavior: TextHeightBehaviorEnumValues;
+
+    // other enums
+    readonly VerticalTextAlign: VerticalTextAlignEnumValues;
+    readonly ResizePolicy: ResizePolicyEnumValues;
 
     // Paragraph Constants
     readonly NoDecoration: number;
@@ -694,6 +699,40 @@ export interface WebGPUCanvasContext {
      * draw multiple frames, e.g. of an animation.
      */
     requestAnimationFrame(drawFrame: (_: Canvas) => void): void;
+}
+
+/**
+ * The glyph and grapheme cluster information associated with a code point within
+ * a paragraph.
+ */
+export interface GlyphInfo {
+    /**
+     * The layout bounds of the grapheme cluster the code point belongs to, in
+     * the paragraph's coordinates.
+     *
+     * This width of the rect is horizontal advance of the grapheme cluster,
+     * the height of the rect is the line height when the grapheme cluster
+     * occupies a full line.
+     */
+    graphemeLayoutBounds: Rect;
+    /**
+     * The left-closed-right-open UTF-16 range of the grapheme cluster the code
+     * point belongs to.
+     */
+    graphemeClusterTextRange: URange;
+    /** The writing direction of the grapheme cluster. */
+    dir: TextDirection;
+    /**
+     * Whether the associated glyph points to an ellipsis added by the text
+     * layout library.
+     *
+     * The text layout library truncates the lines that exceed the specified
+     * max line number, and may add an ellipsis to replace the last few code
+     * points near the logical end of the last visible line. If True, this object
+     * marks the logical end of the list of GlyphInfo objects that are
+     * retrievable from the text layout library.
+     */
+    isEllipsis: boolean;
 }
 
 /**
@@ -941,6 +980,34 @@ export interface TransformProperty {
     value: TransformValue;
 }
 
+/**
+ * Text property for ManagedAnimation's slot support
+ */
+export interface SlottableTextProperty {
+    typeface?: Typeface
+    text?: string;
+
+    textSize?: number;
+    minTextSize?: number;
+    maxTextSize?: number;
+    strokeWidth?: number;
+    lineHeight?: number;
+    lineShift?: number;
+    ascent?: number;
+    maxLines?: number;
+
+    horizAlign?: TextAlignEnumValues;
+    vertAlign?: VerticalTextAlignEnumValues;
+    strokeJoin?: StrokeJoinEnumValues;
+    direction?: TextDirectionEnumValues;
+    linebreak?: LineBreakTypeEnumValues;
+    resize?: ResizePolicyEnumValues;
+
+    boundingBox?: InputRect;
+    fillColor?: InputColor;
+    strokeColor?: InputColor;
+}
+
 export interface ManagedSkottieAnimation extends SkottieAnimation {
     setColor(key: string, color: InputColor): boolean;
     setOpacity(key: string, opacity: number): boolean;
@@ -961,11 +1028,13 @@ export interface ManagedSkottieAnimation extends SkottieAnimation {
     setColorSlot(key: string, color: InputColor): boolean;
     setScalarSlot(key: string, scalar: number): boolean;
     setVec2Slot(key: string, vec2: InputVector2): boolean;
+    setTextSlot(key: string, text: SlottableTextProperty): boolean;
     setImageSlot(key: string, assetName: string): boolean;
 
     getColorSlot(key: string): Color | null;
     getScalarSlot(key: string): number | null;
     getVec2Slot(key: string): Vector2 | null;
+    getTextSlot(key: string): SlottableTextProperty
 
     // Attach a WYSIWYG editor to the text layer identified by 'id' and 'index' (multiple layers
     // can be grouped with the same ID).
@@ -996,14 +1065,42 @@ export interface Paragraph extends EmbindObject<"Paragraph"> {
      * with the top left corner as the origin, and +y direction as down.
      */
     getGlyphPositionAtCoordinate(dx: number, dy: number): PositionWithAffinity;
+    /**
+     * Returns the information associated with the closest glyph at the specified
+     * paragraph coordinate, or null if the paragraph is empty.
+     */
+    getClosestGlyphInfoAtCoordinate(dx: number, dy: number): GlyphInfo | null;
+    /**
+     * Returns the information associated with the glyph at the specified UTF-16
+     * offset within the paragraph's visible lines, or null if the index is out
+     * of bounds, or points to a codepoint that is logically after the last
+     * visible codepoint.
+     */
+    getGlyphInfoAt(index: number): GlyphInfo | null;
 
     getHeight(): number;
     getIdeographicBaseline(): number;
+    /**
+     * Returns the line number of the line that contains the specified UTF-16
+     * offset within the paragraph, or -1 if the index is out of bounds, or
+     * points to a codepoint that is logically after the last visible codepoint.
+     */
+    getLineNumberAt(index: number): number;
     getLineMetrics(): LineMetrics[];
+    /**
+     * Returns the LineMetrics of the line at the specified line number, or null
+     * if the line number is out of bounds, or is larger than or equal to the
+     * specified max line number.
+     */
+    getLineMetricsAt(lineNumber: number): LineMetrics | null;
     getLongestLine(): number;
     getMaxIntrinsicWidth(): number;
     getMaxWidth(): number;
     getMinIntrinsicWidth(): number;
+    /**
+     * Returns the total number of visible lines in the paragraph.
+     */
+    getNumberOfLines(): number;
     getRectsForPlaceholders(): RectWithDirection[];
 
     /**
@@ -4088,6 +4185,15 @@ export interface TextStyleConstructor {
     new(ts: TextStyle): TextStyle;
 }
 
+export interface SlottableTextPropertyConstructor {
+   /**
+     * Fills out all optional fields with defaults. The emscripten bindings complain if there
+     * is a field undefined and it was expecting a float (for example).
+     * @param text
+     */
+   new(text: SlottableTextProperty): SlottableTextProperty;
+}
+
 export interface TypefaceFactory {
     /**
      * Create a typeface using Freetype from the specified bytes and return it. CanvasKit supports
@@ -4717,4 +4823,30 @@ export interface ModifierKeyEnumValues extends EmbindEnum {
     Option: ModifierKey;
     Command: ModifierKey;
     FirstPress: ModifierKey;
+}
+
+export type VerticalAlign = EmbindEnumEntity;
+
+export interface VerticalTextAlignEnumValues extends EmbindEnum {
+    Top: VerticalAlign;
+    TopBaseline: VerticalAlign;
+
+    // Skottie vertical alignment extensions
+    // Visual alignement modes -- these are using tight visual bounds for the paragraph.
+    VisualTop: VerticalAlign;     // visual top    -> text box top
+    VisualCenter: VerticalAlign;  // visual center -> text box center
+    VisualBottom: VerticalAlign;  // visual bottom -> text box bottom
+}
+
+export type ResizePolicy = EmbindEnumEntity;
+
+export interface ResizePolicyEnumValues extends EmbindEnum {
+    // Use the specified text size.
+    None: ResizePolicy;
+    // Resize the text such that the extent box fits (snuggly) in the text box,
+    // both horizontally and vertically.
+    ScaleToFit: ResizePolicy;
+    // Same kScaleToFit if the text doesn't fit at the specified font size.
+    // Otherwise, same as kNone.
+    DownscaleToFit: ResizePolicy;
 }