+++ tests/SkSLTest.cpp
@@ -57,6 +57,7 @@
 #include <cstddef>
 #include <cstdint>
 #include <memory>
+#include <regex>
 #include <string>
 #include <string_view>
 #include <vector>
@@ -68,6 +69,7 @@
 #include "include/gpu/graphite/Surface.h"
 #include "src/gpu/graphite/Caps.h"
 #include "src/gpu/graphite/ContextPriv.h"
+#include "tools/graphite/GraphiteTestContext.h"
 #if defined(SK_DAWN)
 #include "src/gpu/graphite/dawn/DawnCaps.h"
 #endif
@@ -236,13 +238,297 @@ static bool failure_is_expected(std::string_view deviceName,    // "Geforce RTX4
                                 std::string_view backendAPI,    // "OpenGL"
                                 std::string_view name,          // "MatrixToVectorCast"
                                 skiatest::TestType testType) {  // skiatest::TestType::kGraphite
-    // TODO(b/40044139): migrate test-disable list from dm_flags into this file
+    enum TestTypeMatcher { CPU, Ganesh, Graphite, GPU /* either Ganesh or Graphite */ };
+
+    struct TestDisable {
+        std::optional<std::regex> deviceName;
+        std::optional<std::string_view> backendAPI;
+        std::optional<TestTypeMatcher> testTypeMatcher;
+        std::optional<bool> platform;
+    };
+
+    using TestDisableMap = THashMap<std::string_view, std::vector<TestDisable>>;
+
+    // TODO(b/40044139): migrate test-disable list from dm_flags into this map
+    static SkNoDestructor<TestDisableMap> testDisableMap{[] {
+        #define ADRENO "Adreno \\(TM\\) "
+        #define NVIDIA "(Tegra|Quadro|RTX|GTX) "
+
+        TestDisableMap disables;
+        constexpr std::nullopt_t _ = std::nullopt;
+        using regex = std::regex;
+
+#if defined(SK_BUILD_FOR_UNIX)
+        constexpr bool kLinux = true;
+#else
+        constexpr bool kLinux = false;
+#endif
+#if defined(SK_BUILD_FOR_MAC)
+        constexpr bool kMac = true;
+#else
+        constexpr bool kMac = false;
+#endif
+#if defined(SK_BUILD_FOR_IOS)
+        constexpr bool kiOS = true;
+#else
+        constexpr bool kiOS = false;
+#endif
+#if defined(SK_BUILD_FOR_WIN)
+        constexpr bool kWindows = true;
+#else
+        constexpr bool kWindows = false;
+#endif
+#if defined(SK_BUILD_FOR_ANDROID)
+        constexpr bool kAndroid = true;
+#else
+        constexpr bool kAndroid = false;
+#endif
+
+        // - Apple --------------------------------------------------------------------------------
+        // MacOS/iOS do not handle short-circuit evaluation properly in OpenGL (chromium:307751)
+        for (const char* test : {"LogicalAndShortCircuit",
+                                 "LogicalOrShortCircuit"}) {
+            disables[test].push_back({_, "OpenGL", GPU, kMac || kiOS});
+        }
+
+        // ANGLE has a handful of Mac-specific bugs.
+        for (const char* test : {"MatrixScalarNoOpFolding",         // anglebug.com/7525
+                                 "MatrixScalarMath",                // anglebug.com/7525
+                                 "SwizzleIndexStore",               // Apple bug FB12055941
+                                 "OutParamsAreDistinctFromGlobal",  // anglebug.com/7145
+                                 "IntrinsicMixFloatES3"}) {         // anglebug.com/7245
+            disables[test].push_back({_, "ANGLE", GPU, kMac});
+        }
+
+        // Switch fallthrough has some issues on iOS.
+        disables["SwitchWithFallthrough"].push_back({_, "OpenGL", GPU, kiOS});
+
+        // Pack intrinsics do not work properly on Mac + OpenGL + Radeon. (b/40043515)
+        disables["SwitchWithFallthrough"].push_back({regex("Radeon"), "OpenGL", GPU, kMac});
+
+        // - ARM ----------------------------------------------------------------------------------
+        // Mali 400 is a very old driver its share of quirks, particularly in relation to matrices.
+        for (const char* test : {"Matrices",                // b/40043539
+                                 "MatrixNoOpFolding",
+                                 "MatrixScalarMath",        // b/40043764
+                                 "MatrixSwizzleStore",
+                                 "MatrixScalarNoOpFolding", // b/40044644
+                                 "UnaryPositiveNegative",
+                                 "Cross"}) {
+            disables[test].push_back({regex("Mali-400"), _, GPU, _});
+        }
+
+        // - Nvidia -------------------------------------------------------------------------------
+        // Tegra3 has several issues, but the inability to break from a for loop is a common theme.
+        for (const char* test : {"Switch",                            // b/40043561
+                                 "SwitchDefaultOnly",                 //  "      "
+                                 "SwitchWithFallthrough",             //  "      "
+                                 "SwitchWithFallthroughAndVarDecls",  //  "      "
+                                 "SwitchWithLoops",                   //  "      "
+                                 "SwitchCaseFolding",                 //  "      "
+                                 "LoopFloat",                         //  "      "
+                                 "LoopInt",                           //  "      "
+                                 "MatrixScalarNoOpFolding",           // b/40044644
+                                 "MatrixScalarMath",                  // b/40043764
+                                 "MatrixFoldingES2",                  // b/40043017
+                                 "MatrixEquality",                    // b/40043017
+                                 "IntrinsicFract",
+                                 "ModifiedStructParametersCannotBeInlined"}) {
+            disables[test].push_back({regex("Tegra 3"), _, GPU, _});
+        }
+
+        // Various Nvidia GPUs generate errors when assembling weird matrices, and erroneously
+        // constant-fold expressions with side-effects in constructors when compiling GLSL.
+        for (const char* test : {"MatrixConstructorsES2",    // b/40043524
+                                 "MatrixConstructorsES3",    // b/40043524
+                                 "MatrixScalarNoOpFolding",  // b/40044644
+                                 "PreserveSideEffects",      // b/40044140
+                                 "StructFieldNoFolding"}) {  // b/40044479
+            disables[test].push_back({regex(NVIDIA), "OpenGL", _, _});
+            disables[test].push_back({regex(NVIDIA), "ANGLE GL", _, _});
+        }
+
+        disables["IntrinsicMixFloatES3"].push_back({regex("RTX "), "Vulkan", _, kWindows});
+
+        // The Golo features P400s with older and buggier drivers than usual.
+        for (const char* test : {"PreserveSideEffects",  // b/40044140
+                                 "CommaSideEffects"}) {
+            disables[test].push_back({regex("Quadro P400"), _, _, kLinux});
+        }
+
+        // - PowerVR ------------------------------------------------------------------------------
+        for (const char* test : {"OutParamsAreDistinct"              // b/40044222
+                                 "OutParamsAreDistinctFromGlobal"}) {
+            disables[test].push_back({regex("PowerVR Rogue GE8300"), _, GPU, _});
+        }
+
+        // - Radeon -------------------------------------------------------------------------------
+        for (const char* test : {"IntrinsicAll",               // b/40045114
+                                 "MatrixConstructors",         // b/40043524
+                                 "MatrixScalarNoOpFolding",    // b/40044644
+                                 "StructIndexStore",           // b/40045236
+                                 "SwizzleIndexLookup",         // b/40045254
+                                 "SwizzleIndexStore"}) {       // b/40045254
+            disables[test].push_back({regex("Radeon.*(R9|HD)"), "OpenGL", GPU, _});
+            disables[test].push_back({regex("Radeon.*(R9|HD)"), "ANGLE GL", GPU, _});
+        }
+
+        // The Radeon Vega 6 doesn't return zero for the derivative of a uniform.
+        for (const char* test : {"IntrinsicDFdy",
+                                 "IntrinsicDFdx",
+                                 "IntrinsicFwidth"}) {
+            disables[test].push_back({regex("AMD RADV RENOIR"), _, GPU, _});
+        }
+
+        // - Adreno -------------------------------------------------------------------------------
+        // Disable broken tests on Android with Adreno GPUs (b/40043413, b/40045254)
+        for (const char* test : {"ArrayCast",
+                                 "ArrayComparison",
+                                 "CommaSideEffects",
+                                 "IntrinsicMixFloatES2",
+                                 "IntrinsicClampFloat",
+                                 "SwitchWithFallthrough",
+                                 "SwizzleIndexLookup",
+                                 "SwizzleIndexStore"}) {
+            disables[test].push_back({regex(ADRENO "[3456]"), _, _, kAndroid});
+        }
+
+        // Older Adreno 5/6xx drivers report a pipeline error or silently fail when handling inouts.
+        for (const char* test : {"VoidInSequenceExpressions",  // b/295217166
+                                 "InoutParameters",            // b/40043966
+                                 "OutParams",
+                                 "OutParamsDoubleSwizzle",
+                                 "OutParamsNoInline",
+                                 "OutParamsFunctionCallInArgument"}) {
+            disables[test].push_back({regex(ADRENO "[56]"), "Vulkan", _, kAndroid});
+        }
+
+        for (const char* test : {"MatrixToVectorCast",     // b/40043288
+                                 "StructsInFunctions"}) {  // b/40043024
+            disables[test].push_back({regex(ADRENO "[345]"), "OpenGL", _, kAndroid});
+        }
+
+        // Constructing a matrix from vectors and scalars can be surprisingly finicky (b/40043539)
+        for (const char* test : {"Matrices",
+                                 "MatrixNoOpFolding"}) {
+            disables[test].push_back({regex(ADRENO "3"), "OpenGL", _, kAndroid});
+        }
+
+        // Adreno 600 doesn't handle isinf() in OpenGL correctly. (b/40043464)
+        disables["IntrinsicIsInf"].push_back({regex(ADRENO "6"), "OpenGL", _, kAndroid});
+
+        // Older Adreno drivers crash when presented with an empty block (b/40044390)
+        disables["EmptyBlocksES3"].push_back({regex(ADRENO "(540|630)"), _, _, kAndroid});
+
+        // Adrenos alias out-params to globals improperly (b/40044222)
+        disables["OutParamsAreDistinctFromGlobal"].push_back({regex(ADRENO "[3456]"), "OpenGL",
+                                                              _, kAndroid});
+        // Adreno generates the wrong result for this test. (b/40044477)
+        disables["StructFieldFolding"].push_back({regex(ADRENO "[56]"), "OpenGL",
+                                                        _, kAndroid});
+
+        // - Intel --------------------------------------------------------------------------------
+        // Disable various tests on Intel.
+        // Intrinsic floor() on Intel + ANGLE + DirectX is broken (anglebug.com/5588)
+        disables["IntrinsicFloor"].push_back({regex("Intel.*(Iris|HD)"), "ANGLE D3D", _, _});
+
+        // Intrinsic not() and mix() are broken on Intel GPUs in Metal. (b/40045105)
+        for (const char* test : {"IntrinsicNot",
+                                 "IntrinsicMixFloatES3"}) {
+            disables[test].push_back({regex("Intel.*(Iris|6000)"), "Metal", _, kMac});
+        }
+
+        // Swizzled-index store is broken across many Intel GPUs. (b/40045254)
+        disables["SwizzleIndexStore"].push_back({regex("Intel"), "OpenGL", _, kMac});
+        disables["SwizzleIndexStore"].push_back({regex("Intel.*Iris"), _, _, kWindows});
+
+        // vec4(mat2) conversions can lead to a crash on Intel + ANGLE (b/40043275)
+        for (const char* test : {"VectorToMatrixCast",
+                                 "VectorScalarMath",
+                                 "TrivialArgumentsInlineDirectly"}) {
+            disables[test].push_back({regex("Intel"), "ANGLE", _, kWindows});
+        }
+
+        for (const char* test : {"MatrixFoldingES2",
+                                 "MatrixEquality",
+                                 "TemporaryIndexLookup",  // b/40045228
+                                 "SwizzleIndexLookup"}) { // b/40045254
+            disables[test].push_back({regex("Intel.*(Iris|4400)"), "OpenGL", _, kWindows});
+            disables[test].push_back({regex("Intel.*(Iris|4400)"), "ANGLE",  _, kWindows});
+        }
+
+        for (const char* test : {"ReturnsValueOnEveryPathES3",      // b/40043548
+                                 "OutParamsAreDistinctFromGlobal",  // b/40044222
+                                 "StructFieldFolding"}) {           // b/40044477
+            disables[test].push_back({regex("Intel"), "OpenGL", _, kWindows});
+            disables[test].push_back({regex("Intel"), "ANGLE GL", _, kWindows});
+        }
+
+        for (const char* test : {"SwitchDefaultOnly",               // b/40043548
+                                 "ReturnsValueOnEveryPathES3"}) {   // b/40045205
+            disables[test].push_back({regex("Intel"), "Vulkan", _, kLinux});
+        }
+
+        for (const char* test : {"SwitchDefaultOnly"}) {
+            disables[test].push_back({regex("Intel"), "ANGLE", _, kWindows});
+        }
+
+        for (const char* test : {"SwizzleAsLValueES3"}) {  // https://anglebug.com/8260
+            disables[test].push_back({regex("Intel"), _, _, kWindows});
+            disables[test].push_back({_, "ANGLE", _, kWindows});
+        }
+
+        // Some Intel GPUs don't return zero for the derivative of a uniform.
+        for (const char* test : {"IntrinsicDFdy",
+                                 "IntrinsicDFdx",
+                                 "IntrinsicFwidth"}) {
+            disables[test].push_back({regex("Intel"), _, GPU, _});
+        }
+
+        disables["LoopFloat"].push_back({regex("Intel.*(Iris|6000)"), _, _, kMac});  // b/40043507
+
+        #undef ADRENO
+        #undef NVIDIA
+
+        return disables;
+    }()};
+
+    if (const std::vector<TestDisable>* testDisables = testDisableMap->find(name)) {
+        for (const TestDisable& d : *testDisables) {
+            if (d.platform.has_value() && !*d.platform) {
+                continue;  // disable applies to a different platform
+            }
+            if (d.backendAPI.has_value() && !skstd::contains(backendAPI, *d.backendAPI)) {
+                continue;  // disable applies to a different backend API
+            }
+            if (d.deviceName.has_value() &&
+                !std::regex_search(deviceName.begin(), deviceName.end(), *d.deviceName)) {
+                continue;  // disable applies to a different device
+            }
+            if (d.testTypeMatcher == CPU && testType != skiatest::TestType::kCPU) {
+                continue;  // disable only applies to CPU
+            }
+            if (d.testTypeMatcher == Ganesh && testType != skiatest::TestType::kGanesh) {
+                continue;  // disable only applies to Ganesh
+            }
+            if (d.testTypeMatcher == Graphite && testType != skiatest::TestType::kGraphite) {
+                continue;  // disable only applies to Graphites
+            }
+            if (d.testTypeMatcher == GPU && testType == skiatest::TestType::kCPU) {
+                continue;  // disable only applies to GPU
+            }
+            // This test was disabled.
+            return true;
+        }
+    }
+
+    // This test was not in our disable list.
     return false;
 }
 
 static void test_one_permutation(skiatest::Reporter* r,
                                  std::string_view deviceName,
-                                 std::string_view backend,
+                                 std::string_view backendAPI,
                                  SkSurface* surface,
                                  const char* name,
                                  const char* testFile,
@@ -256,6 +542,13 @@ static void test_one_permutation(skiatest::Reporter* r,
     SkRuntimeEffect::Result result = SkRuntimeEffect::MakeForShader(shaderString, options);
     if (!result.effect) {
         ERRORF(r, "%s%s: %s", testFile, permutationSuffix, result.errorText.c_str());
+        return;
+    }
+    if (failure_is_expected(deviceName, backendAPI, name, testType)) {
+        // Some driver bugs can be catastrophic (e.g. crashing dm entirely), so we don't even try to
+        // run a shader if we expect that it might fail.
+        SkDebugf("%s: skipped %.*s%s\n", testFile, (int)backendAPI.size(), backendAPI.data(),
+                                         permutationSuffix);
         return;
     }
 
@@ -287,7 +580,7 @@ static void test_one_permutation(skiatest::Reporter* r,
                                           "%02X%02X%02X%02X %02X%02X%02X%02X",
                                           permutationSuffix,
                                           (int)deviceName.size(), deviceName.data(),
-                                          (int)backend.size(),    backend.data(),
+                                          (int)backendAPI.size(), backendAPI.data(),
 
                                           SkColorGetR(color[0][0]), SkColorGetG(color[0][0]),
                                           SkColorGetB(color[0][0]), SkColorGetA(color[0][0]),
@@ -301,17 +594,13 @@ static void test_one_permutation(skiatest::Reporter* r,
                                           SkColorGetR(color[1][1]), SkColorGetG(color[1][1]),
                                           SkColorGetB(color[1][1]), SkColorGetA(color[1][1]));
 
-        if (failure_is_expected(deviceName, backend, name, testType)) {
-            INFOF(r, "(KNOWN ISSUE) %s", message.c_str());
-        } else {
-            ERRORF(r, "%s", message.c_str());
-        }
+        ERRORF(r, "%s", message.c_str());
     }
 }
 
 static void test_permutations(skiatest::Reporter* r,
                               std::string_view deviceName,
-                              std::string_view backend,
+                              std::string_view backendAPI,
                               SkSurface* surface,
                               const char* name,
                               const char* testFile,
@@ -320,10 +609,10 @@ static void test_permutations(skiatest::Reporter* r,
     SkRuntimeEffect::Options options = strictES2 ? SkRuntimeEffect::Options{}
                                                  : SkRuntimeEffectPriv::ES3Options();
     options.forceUnoptimized = false;
-    test_one_permutation(r, deviceName, backend, surface, name, testFile, testType, "", options);
+    test_one_permutation(r, deviceName, backendAPI, surface, name, testFile, testType, "", options);
 
     options.forceUnoptimized = true;
-    test_one_permutation(r, deviceName, backend, surface, name, testFile, testType,
+    test_one_permutation(r, deviceName, backendAPI, surface, name, testFile, testType,
                          " (Unoptimized)", options);
 }
 
@@ -385,6 +674,7 @@ static void test_ganesh(skiatest::Reporter* r,
 #if defined(SK_GRAPHITE)
 static void test_graphite(skiatest::Reporter* r,
                           skgpu::graphite::Context* ctx,
+                          skiatest::graphite::GraphiteTestContext* testCtx,
                           const char* name,
                           const char* testFile,
                           SkSLTestFlags flags) {
@@ -415,14 +705,7 @@ static void test_graphite(skiatest::Reporter* r,
                                                 kPremul_SkAlphaType);
     sk_sp<SkSurface> surface = SkSurfaces::RenderTarget(recorder.get(), info);
     std::string_view deviceName = ctx->priv().caps()->deviceName();
-    std::string_view backendAPI;
-    switch (ctx->backend()) {
-        case skgpu::BackendApi::kDawn:   backendAPI = "Dawn"; break;
-        case skgpu::BackendApi::kMetal:  backendAPI = "Metal"; break;
-        case skgpu::BackendApi::kVulkan: backendAPI = "Vulkan"; break;
-        case skgpu::BackendApi::kMock:   backendAPI = "Mock"; break;
-        default:                         backendAPI = "Unknown"; break;
-    }
+    std::string_view backendAPI = skgpu::ContextTypeName(testCtx->contextType());
 
     if (shouldRunGPU) {
         test_permutations(r, deviceName, backendAPI, surface.get(), name, testFile,
@@ -609,10 +892,11 @@ static bool is_native_context_or_dawn(skgpu::ContextType type) {
                                                is_native_context_or_dawn, \
                                                r,                         \
                                                context,                   \
+                                               testContext,               \
                                                /*opt_filter=*/nullptr,    \
                                                is_gpu(flags),             \
                                                ctsEnforcement) {          \
-        test_graphite(r, context, #name, path, flags);                    \
+        test_graphite(r, context, testContext, #name, path, flags);       \
     }
 #else
 #define DEF_GRAPHITE_SKSL_TEST(flags, ctsEnforcement, name, path) /* Graphite is disabled */
@@ -664,9 +948,7 @@ SKSL_TEST(CPU | GPU,     kApiLevel_U, MatrixNoOpFolding,               "folding/
 SKSL_TEST(CPU | GPU,     kApiLevel_U, MatrixScalarNoOpFolding,         "folding/MatrixScalarNoOpFolding.rts")
 SKSL_TEST(CPU | GPU,     kApiLevel_U, MatrixVectorNoOpFolding,         "folding/MatrixVectorNoOpFolding.rts")
 SKSL_TEST(CPU | GPU,     kApiLevel_T, Negation,                        "folding/Negation.rts")
-// TODO(skia:13035): This test fails on Nvidia GPUs on OpenGL but passes Vulkan. Re-enable the test
-// on Vulkan when granular GPU backend selection is supported.
-SKSL_TEST(CPU,           kApiLevel_T, PreserveSideEffects,             "folding/PreserveSideEffects.rts")
+SKSL_TEST(CPU | GPU,     kApiLevel_T, PreserveSideEffects,             "folding/PreserveSideEffects.rts")
 SKSL_TEST(CPU | GPU,     kApiLevel_T, SelfAssignment,                  "folding/SelfAssignment.rts")
 SKSL_TEST(CPU | GPU,     kApiLevel_T, ShortCircuitBoolFolding,         "folding/ShortCircuitBoolFolding.rts")
 SKSL_TEST(CPU | GPU,     kApiLevel_U, StructFieldFolding,              "folding/StructFieldFolding.rts")
@@ -700,8 +982,7 @@ SKSL_TEST(CPU | GPU,     kApiLevel_T, InlineWithModifiedArgument,               
 SKSL_TEST(CPU | GPU,     kApiLevel_T, InlineWithNestedBigCalls,                         "inliner/InlineWithNestedBigCalls.sksl")
 SKSL_TEST(CPU | GPU,     kApiLevel_T, InlineWithUnmodifiedArgument,                     "inliner/InlineWithUnmodifiedArgument.sksl")
 SKSL_TEST(CPU | GPU,     kApiLevel_T, InlineWithUnnecessaryBlocks,                      "inliner/InlineWithUnnecessaryBlocks.sksl")
-// TODO(tint:1932): GPU is disabled for now; test exposes a bug in the Tint SPIR-V Reader
-SKSL_TEST(CPU,           kNextRelease,IntrinsicNameCollision,                           "inliner/IntrinsicNameCollision.sksl")
+SKSL_TEST(CPU | GPU,     kNextRelease,IntrinsicNameCollision,                           "inliner/IntrinsicNameCollision.sksl")
 SKSL_TEST(CPU | GPU,     kNextRelease,ModifiedArrayParametersCannotBeInlined,           "inliner/ModifiedArrayParametersCannotBeInlined.sksl")
 SKSL_TEST(CPU | GPU,     kNextRelease,ModifiedStructParametersCannotBeInlined,          "inliner/ModifiedStructParametersCannotBeInlined.sksl")
 SKSL_TEST(CPU | GPU,     kApiLevel_T, NoInline,                                         "inliner/NoInline.sksl")
@@ -751,8 +1032,7 @@ SKSL_TEST(ES3 | GPU_ES3, kNever,      IntrinsicMixFloatES3,            "intrinsi
 SKSL_TEST(GPU_ES3,       kNever,      IntrinsicModf,                   "intrinsics/Modf.sksl")
 SKSL_TEST(CPU | GPU,     kNever,      IntrinsicNot,                    "intrinsics/Not.sksl")
 SKSL_TEST(GPU_ES3,       kNever,      IntrinsicOuterProduct,           "intrinsics/OuterProduct.sksl")
-// Fails on Mac OpenGL + Radeon 5300M (skia:12434)
-// SKSL_TEST(GPU_ES3,    kNever,      IntrinsicPackUnorm2x16,          "intrinsics/PackUnorm2x16.sksl")
+SKSL_TEST(GPU_ES3,       kNever,      IntrinsicPackUnorm2x16,          "intrinsics/PackUnorm2x16.sksl")
 SKSL_TEST(CPU | GPU,     kNever,      IntrinsicRadians,                "intrinsics/Radians.sksl")
 SKSL_TEST(GPU_ES3,       kNever,      IntrinsicRound,                  "intrinsics/Round.sksl")
 SKSL_TEST(GPU_ES3,       kNever,      IntrinsicRoundEven,              "intrinsics/RoundEven.sksl")
@@ -798,8 +1078,7 @@ SKSL_TEST(CPU | GPU,     kNever,      DeadGlobals,                     "shared/D
 SKSL_TEST(ES3 | GPU_ES3, kNever,      DeadLoopVariable,                "shared/DeadLoopVariable.sksl")
 SKSL_TEST(CPU | GPU,     kApiLevel_T, DeadIfStatement,                 "shared/DeadIfStatement.sksl")
 SKSL_TEST(CPU | GPU,     kApiLevel_T, DeadReturn,                      "shared/DeadReturn.sksl")
-// TODO(skia:12012): some Radeons crash when compiling this code; disable them.
-SKSL_TEST(ES3/* | GPU_ES3*/,kNever,   DeadReturnES3,                   "shared/DeadReturnES3.sksl")
+SKSL_TEST(ES3 | GPU_ES3, kNever,      DeadReturnES3,                   "shared/DeadReturnES3.sksl")
 SKSL_TEST(CPU | GPU,     kApiLevel_T, DeadStripFunctions,              "shared/DeadStripFunctions.sksl")
 SKSL_TEST(CPU | GPU,     kApiLevel_T, DependentInitializers,           "shared/DependentInitializers.sksl")
 SKSL_TEST(CPU | GPU,     kApiLevel_U, DoubleNegation,                  "shared/DoubleNegation.sksl")
@@ -824,10 +1103,8 @@ SKSL_TEST(CPU | GPU,     kApiLevel_U, LogicalAndShortCircuit,          "shared/L
 SKSL_TEST(CPU | GPU,     kApiLevel_U, LogicalOrShortCircuit,           "shared/LogicalOrShortCircuit.sksl")
 SKSL_TEST(CPU | GPU,     kApiLevel_T, Matrices,                        "shared/Matrices.sksl")
 SKSL_TEST(ES3 | GPU_ES3, kNever,      MatricesNonsquare,               "shared/MatricesNonsquare.sksl")
-// TODO(skia:12443) The MatrixConstructors tests actually don't work on MANY devices. The GLSL SkQP
-// suite does a terrible job of enforcing this rule. We still test the ES2 variant on CPU.
-SKSL_TEST(CPU,           kNever,      MatrixConstructorsES2,           "shared/MatrixConstructorsES2.sksl")
-SKSL_TEST(ES3,           kNever,      MatrixConstructorsES3,           "shared/MatrixConstructorsES3.sksl")
+SKSL_TEST(CPU | GPU,     kNever,      MatrixConstructorsES2,           "shared/MatrixConstructorsES2.sksl")
+SKSL_TEST(ES3 | GPU_ES3, kNever,      MatrixConstructorsES3,           "shared/MatrixConstructorsES3.sksl")
 SKSL_TEST(CPU | GPU,     kApiLevel_T, MatrixEquality,                  "shared/MatrixEquality.sksl")
 SKSL_TEST(CPU | GPU,     kNextRelease,MatrixIndexLookup,               "shared/MatrixIndexLookup.sksl")
 SKSL_TEST(CPU | GPU,     kNextRelease,MatrixIndexStore,                "shared/MatrixIndexStore.sksl")