+++ modules/canvaskit/CHANGELOG.md
@@ -19,10 +19,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Blender.Mode` Create a blender that implements the specified BlendMode.
 - `RuntimeEffect.MakeForBlender` Compiles a RuntimeEffect from the given blender code.
 - `ManagedAnimation` getters and setters for lottie slots exported by Essential Graphics in AE.
-   Color, scalar, vec2, and image slot types are supported.
+   Color, scalar, vec2, text, and image slot types are supported.
 - `ManagedAnimation` WYSIWYG editor API: `attachEditor`, `enableEditor`, `dispatchEditorKey`,
   `dispatchEditorPointer`.
 - `InputState` and `ModifierKey` enums.
+- `Paragraph.getClosestGlyphInfoAtCoordinate` and `Paragraph.getGlyphInfoAt` return the information associated with the glyph or grapheme cluster in the paragraph at the specified location/index.
+- `Paragraph.getLineMetricsAt`, returns the line metrics of a line.
+- `Paragraph.getNumberOfLines`, returns the number of visible lines in the paragraph.
+- `Paragraph.getLineNumberAt`, finds the line that contains the given UTF-16 index.
+
 
 ### Fixed
  - `EmbindObject` has been updated to allow TypeScript to differentiate between opaque